requirements.txt
bcolz==1.2.1
bokeh @ file:///tmp/build/80754af9/bokeh_1638362823963/work
Bottleneck==1.3.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.11.0
dask==2021.10.0
distributed @ file:///tmp/build/80754af9/distributed_1635968201545/work
distro==1.6.0
easydict==1.7
Flask==2.0.2
Flask-Uploads @ git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc93e684fa0a3a4350a38e41ae00483a796
fsspec @ file:///opt/conda/conda-bld/fsspec_1642510437511/work
graphviz==0.8.4
HeapDict @ file:///Users/ktietz/demo/mc3/conda-bld/heapdict_1630598515714/work
idna==2.6
importlib-metadata==4.11.0
itsdangerous==2.0.1
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282414123/work
locket==0.2.1
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592846017414/work
mkl-service==2.4.0
mock @ file:///tmp/build/80754af9/mock_1607622725907/work
msgpack @ file:///tmp/build/80754af9/msgpack-python_1612287175283/work
mxnet==1.9.0
numexpr @ file:///tmp/build/80754af9/numexpr_1640686084435/work
numpy==1.21.5
olefile==0.46
opencv-python==4.5.5.62
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
partd @ file:///tmp/build/80754af9/partd_1618000087440/work
Pillow==8.4.0
protobuf==3.19.4
psutil @ file:///tmp/build/80754af9/psutil_1612298016854/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scikit-build==0.12.0
scikit-learn==0.19.2
scipy==1.1.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1623949099177/work
tables==3.6.1
tblib @ file:///Users/ktietz/demo/mc3/conda-bld/tblib_1629402031467/work
tensorboardX==1.2
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.4.0
torchvision==0.2.1
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.23.4
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.22
Werkzeug==2.0.3
zict==2.0.0
zipp==3.7.0
